Downloads/openenv-code-review/openenv-code-review/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-code-review-env"
version = "1.0.0"
description = "Code Review OpenEnv environment — PR diff analysis with planted security vulnerabilities"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "code_review_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["code_review_env", "code_review_env.server"]
package-dir = { "code_review_env" = ".", "code_review_env.server" = "server" }
